+++ cpp/modmesh/CMakeLists.txt
@@ -39,6 +39,7 @@ add_subdirectory(python)
 add_subdirectory(spacetime)
 add_subdirectory(view)
 add_subdirectory(inout)
+add_subdirectory(serialization)
 
 if(USE_PYTEST_HELPER_BINDING)
     add_subdirectory(testhelper)
@@ -49,16 +50,10 @@ set(MODMESH_ROOT_HEADERS
     ${CMAKE_CURRENT_SOURCE_DIR}/base.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/math.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/grid.hpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/SerializableItem.hpp
-    CACHE FILEPATH "" FORCE)
-
-set(MODMESH_ROOT_SOURCES
-    ${CMAKE_CURRENT_SOURCE_DIR}/SerializableItem.cpp
     CACHE FILEPATH "" FORCE)
 
 set(MODMESH_ROOT_FILES
     ${MODMESH_ROOT_HEADERS}
-    ${MODMESH_ROOT_SOURCES}
     CACHE FILEPATH "" FORCE)
 
 if (BUILD_METAL)
@@ -95,6 +90,7 @@ set(MODMESH_TERMINAL_FILES
     ${MODMESH_PYTHON_FILES}
     ${MODMESH_INOUT_FILES}
     ${MODMESH_TESTHELPER_FILES}
+    ${MODMESH_SERIALIZATION_FILES}
     CACHE FILEPATH "" FORCE)
 
 set(MODMESH_GRAPHIC_FILES
