+++ cpp/modmesh/serialization/SerializableItem.cpp
@@ -25,8 +25,8 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-#include <modmesh/SerializableItem.hpp>
-#include <modmesh/base.hpp>
+
+#include <modmesh/serialization/SerializableItem.hpp>
 
 namespace modmesh
 {
